move confirm functionality
